treesitter: install parsers with nix
This fixes weird issues where you might get errors about libstdc++ after
updating your system without running :TSInstall again. Fixes it by just
relying on Nix to do the managing itself!

Unfortunately, for now this still needs work: you can't select which
languages are to be installed, which is why it's off by default. In the
hopefully not-too-distant future, this will be the default, and the
other way will be deprecated.