Added tests for template
lib/default.nix was missing the check file, which is now moved to
test-derivation.nix.
Here it was important to export both functions as the config for
mkTestDerivation does not support imports from a configuration.

the check workflow also got a test for the template so we can keep it up
to date, when making some breaking changes.