tests: Install all efmls-configs tools
This requires to pass the efmls-configs file to the derivation helpers,
if we try to import it from the test file directly it fails.

We then look into the types to find the available languages &
corresponding tools, to enable them all.

We then filter out the broken tools or the unpackaged tools, in order to
detect more easily the introduction of future tools.