plugins/none-ls: fix withArgs type, syntax (#1182)
This was previously null or string, but was changed to null or Lua in
nix-community/nixvim#1169. Internally, `withArgs` is expected to be a
string, but `mkNullOrLua` applies `mkRaw`, leading to a type error when
generating the output, as it is attempting to interpolate a set rather
than a string.

Tested locally, and verified that it fixes my setup.