changed default to true instead of false
I have been struggling to figure out why my telescope didn't have
borders, as it usually does when normally enabled. Then I came across
this option and saw it was set to false by default. This changes it back
to true to be consistent with every other default telescope
configuration.