plugins/rustaceanvim: fix checkhealth error with neotest configuration
Rustaceanvim will throw an error if the configuration is sourced after
initialization. Our globals are defined at the top of the init.lua so
moving the settings to leverage the globals option and setting a
callback function as default to ensure we dont need to worry about
something initializing it before lspOnAttach is available.