docs: introduce a dedicated homepage
Based of the README for now, with a small additional "welcome to the
docs" section.

Perhaps in future, we could re-use the README.md verbatim? Perhaps with
some additional post-processing.

I think we'd want to dramatically reduce its size first though.

By having a dedicated homepage outside of `user-guide`, this fixes #1625