docs: treat internal options as invisible
`internal` options are intended to be implicitly not `visible`. This is
already handled correctly by the nixos tooling, but our custom tooling
did not correctly handle the edge case.

This lead to strange cases where an internal option is not included in
the docs, but its sub-options were still visible.