plugins/treesitter: don't exclude nvim-treesitter from combining by default
Nvim-treesitter's parsers from nixpkgs don't include grammars anymore.
Originally it was added to standalonePlugins as workaround.
If the user has some other plugin containing treesitter queries, this
change can cause a build failure due to collisions. But since it is
easier to add the plugin to standalonePlugins compared to removing it, I
think this should be the default.