ci/docs: fix `base-href` in matrix
For some reason, when updating `base-href` using `+=`, yq updates _all_
entries matching the LHS selection. This means that all non-main
branches get `"base-href": "/nixvim/24.11/"` in the job matrix.

This gets passed in as an overridden attr when building the docs, and is
used to determine which list item represents the currently-being-built
docs; therefore both 25.05 and 24.11 think they are 24.11...

For some reason, while `+=` does this, `=` does not. So switched to
using that.