flake/ci: fix lazy eval of dev partition
If the `ci` options are a set of options rather than a single option,
then the module system cannot lazily override it via `mkForce`.

In practice, this means that the dev partition gets evaluated strictly
when the module system "pushes down" the `ci` definitions to its
options.

Instead, we must ensure that each attribute listed in `partitionedAttrs`
corresponds to an actual option which can be directly defined, with no
"pushing down" required.

Fixes #3532