ci/update-maintainers: drop `issues:write` permission
We requested this permission to be able to create new labels, when
labelling the PR. However that is not necessary here, because a) we
don't apply any labels and b) if we did we could create them manually.

Our app installation doesn't have `issues:write`, so requesting it was
causing the workflow to fail.