plugins/schemastore: use the correct LSP configuration keys for the new lsp module
The old lsp module wrapped (in plugins/lsp/language-servers/default.nix)
all configuration of the jsonls and yamlls servers in `json` and `yaml`
keys respectively, so the schemastore module never needed to add those
keys.

However, the new lsp module passes the configuration as-is instead. This
means that the schemastore module must be the one to wrap the
configuration in the `json` and `yaml` keys. Without this, jsonls does
not get its support enabled at all, and yamlls falls back to the
built-in support.