lib/utils: init `addSubOptions`
Adds `addSubOptions` and `addSubOptionsWith` util functions to extend an
option-type with docs-only sub-options.

These functions use a similar interface to `types.submodule` and
`types.submoduleWith` to extend an existing option-type with additional
docs sub-options, returned by `getSubOptions`.

The sub-options do not affect the normal operation of the type, such as
checking or merging definitions.